chore: abstract reanimated example
