+++ examples/shared/src/examples/Animations/ReanimatedViewAnnotation.tsx
@@ -1,0 +1,5 @@
+import { AbstractReanimatedComponent } from "./AbstractReanimatedComponent";
+
+export const ReanimatedViewAnnotation = () => {
+  return <AbstractReanimatedComponent component="ViewAnnotation" />;
+};
